Refactor navigation structure; consolidate routing logic and enhance accessibility feature definitions
