+++ hooks/useThemeColor.ts
@@ -1,14 +1,9 @@
-/**
- * Learn more about light and dark modes:
- * https://docs.expo.dev/guides/color-schemes/
- */
-
-import { useColorScheme } from '@/hooks/useColorScheme';
-import { Colors } from '@/theme/colors';
+import { Colors, type ColorKeys } from '@/theme/colors';
+import { useColorScheme } from './useColorScheme';
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
-  colorName: keyof typeof Colors.light & keyof typeof Colors.dark,
+  colorName: ColorKeys,
 ) {
   const theme = useColorScheme() ?? 'light';
   const colorFromProps = props[theme];
