+++ navigation/RootStack.tsx
@@ -1,0 +1,10 @@
+import { Stack } from 'expo-router';
+
+export default function RootStack() {
+  return (
+    <Stack>
+      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen name="+not-found" />
+    </Stack>
+  );
+}
