+++ types/accessibility.ts
@@ -1,3 +1,5 @@
+import { LucideIcon } from "lucide-react-native";
+import { FeatureRoute } from "./navigation";
 
 export const AccessibilityLevel = {
     A: 'A',
@@ -7,3 +9,11 @@ export const AccessibilityLevel = {
 } as const;
 
 export type AccessibilityLevel = (typeof AccessibilityLevel)[keyof typeof AccessibilityLevel]; export type BadgeSize = 'small' | 'medium' | 'large';
+
+export type AccessibilityFeature = {
+    key: string;
+    icon: LucideIcon;
+    title: string;
+    description: string;
+    route: FeatureRoute;
+};
