+++ types/navigation.ts
@@ -1,3 +1,4 @@
 // src/types/navigation.ts
 export type TabScreenName = 'index' | 'components' | 'guidelines' | 'settings';
 export type TabScreenTitle = 'Home' | 'Components' | 'Guidelines' | 'Settings';
+export type FeatureRoute = '/components' | '/guidelines' | null;
